Improve multiple object interaction with GUI.
The gui is now fully aware of multiple objects in the scene.
It updates it's settings depending on which object we click on.
I've removed the updates.js file as it was no longer working
with the gui .onChange() methods.
( See gui.js in the createGuiMaterial() function on line 177 )

It's looking really great!